fastapi/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
langchain-chroma
python-dotenv
sentence-transformers
